Rename app and replace icon
