+++ Home/App.config
@@ -1,0 +1,33 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+    <configSections>
+        <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
+            <section name="Home.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
+        </sectionGroup>
+    </configSections>
+    <startup> 
+        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2" />
+    </startup>
+    <userSettings>
+        <Home.Properties.Settings>
+            <setting name="URL" serializeAs="String">
+                <value>http://demo.openhab.org:8080/basicui/app?sitemap=demo</value>
+            </setting>
+            <setting name="Height" serializeAs="String">
+                <value>400</value>
+            </setting>
+            <setting name="Width" serializeAs="String">
+                <value>600</value>
+            </setting>
+            <setting name="Top" serializeAs="String">
+                <value>0</value>
+            </setting>
+            <setting name="Left" serializeAs="String">
+                <value>0</value>
+            </setting>
+            <setting name="ReloadOnOpen" serializeAs="String">
+                <value>False</value>
+            </setting>
+        </Home.Properties.Settings>
+    </userSettings>
+</configuration>
