+++ Home/Home.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29806.167
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Home", "Home.csproj", "{C9E1E4D2-EC65-4AE1-B6DD-304647839BE8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C9E1E4D2-EC65-4AE1-B6DD-304647839BE8}.Debug|x64.ActiveCfg = Debug|x64
+		{C9E1E4D2-EC65-4AE1-B6DD-304647839BE8}.Debug|x64.Build.0 = Debug|x64
+		{C9E1E4D2-EC65-4AE1-B6DD-304647839BE8}.Release|x64.ActiveCfg = Release|x64
+		{C9E1E4D2-EC65-4AE1-B6DD-304647839BE8}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8663B392-6214-4CBC-A52A-AB7B8866A2E8}
+	EndGlobalSection
+EndGlobal
