+++ Home/Properties/StringResources.en-US.xaml
@@ -1,0 +1,13 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:system="clr-namespace:System;assembly=mscorlib">
+    <system:String x:Key="ApplicationName">Home</system:String>
+    <system:String x:Key="MessageDefaultCaption">Home says...</system:String>
+    <system:String x:Key="MessageAppAlreadyRunning" xml:space="preserve">Home app is already running.&#13;You can find its icon in system tray.</system:String>
+    <system:String x:Key="Exit">Exit</system:String>
+    <system:String x:Key="Settings">Settings</system:String>
+    <system:String x:Key="CloseSettings">CLOSE</system:String>
+    <system:String x:Key="RequiredField">This field is required</system:String>
+    <system:String x:Key="URL">URL</system:String>
+    <system:String x:Key="ReloadOnWindowOpen">Reload page on every window open</system:String>
+</ResourceDictionary>
