+++ Home/Properties/StringResources.ru-RU.xaml
@@ -1,0 +1,13 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:system="clr-namespace:System;assembly=mscorlib">
+    <system:String x:Key="ApplicationName">Home</system:String>
+    <system:String x:Key="MessageDefaultCaption">Home сообщает...</system:String>
+    <system:String x:Key="MessageAppAlreadyRunning" xml:space="preserve">Приложение Home уже запущено.&#13;Вы можете найти иконку приложения в системном лотке рядом с часами.</system:String>
+    <system:String x:Key="Exit">Выход</system:String>
+    <system:String x:Key="Settings">Настройки</system:String>
+    <system:String x:Key="CloseSettings">ЗАКРЫТЬ</system:String>
+    <system:String x:Key="RequiredField">Это обязательное поле</system:String>
+    <system:String x:Key="URL">URL</system:String>
+    <system:String x:Key="ReloadOnWindowOpen">Обновлять страницу при каждом открытии окна</system:String>
+</ResourceDictionary>
