+++ Home/ViewModel/Converter/BooleanToVisibilityConverter.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using System.Globalization;
+using System.Windows;
+using System.Windows.Data;
+
+namespace Home.ViewModel.Converter
+{
+    [ValueConversion(typeof(bool), typeof(Visibility))]
+    public class BooleanToVisibilityConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            var visibility = Visibility.Visible;
+
+            try
+            {
+                if (value != null)
+                {
+                    visibility = (bool)value ? Visibility.Visible : Visibility.Collapsed;
+                }
+            }
+            catch
+            {
+                // ignored
+            }
+
+            return visibility;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return (value != null) && (value == (object)Visibility.Visible);
+        }
+    }
+}
