+++ Home/ViewModel/SettingBindingExtension.cs
@@ -1,0 +1,23 @@
+﻿using System.Windows.Data;
+
+namespace Home.ViewModel
+{
+    public class SettingBindingExtension : Binding
+    {
+        public SettingBindingExtension()
+        {
+            Initialize();
+        }
+
+        public SettingBindingExtension(string path) : base(path)
+        {
+            Initialize();
+        }
+
+        private void Initialize()
+        {
+            Source = Properties.Settings.Default;
+            Mode = BindingMode.TwoWay;
+        }
+    }
+}
