--- TheGate/View/Dialog/SettingsViewDialog.xaml
@@ -1,42 +1,0 @@
-﻿<UserControl x:Class="TheGate.View.Dialog.SettingsViewDialog"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:rule="clr-namespace:TheGate.ViewModel.Rule"
-             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-             xmlns:wpf="clr-namespace:MaterialDesignThemes.Wpf;assembly=MaterialDesignThemes.Wpf"
-             mc:Ignorable="d" 
-             d:DesignHeight="200" d:DesignWidth="200">
-
-    <materialDesign:Card MinWidth="200" Background="{DynamicResource MaterialDesignPaper}">
-        <Grid>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-            </Grid.RowDefinitions>
-
-            <TextBlock Grid.Row="0" Margin="8" Style="{DynamicResource MaterialDesignCaptionTextBlock}" FontSize="18" Text="{DynamicResource Settings}" />
-
-            <TextBox Grid.Row="1" Margin="16" FontSize="18" wpf:HintAssist.Hint="{DynamicResource URL}" Style="{DynamicResource MaterialDesignFloatingHintTextBox}">
-                <TextBox.Text>
-                    <Binding Path="URL" UpdateSourceTrigger="PropertyChanged">
-                        <Binding.ValidationRules>
-                            <rule:RequiredFieldValidationRule ValidatesOnTargetUpdated="True" />
-                        </Binding.ValidationRules>
-                    </Binding>
-                </TextBox.Text>
-            </TextBox>
-
-            <StackPanel Grid.Row="2" Orientation="Horizontal">
-                <TextBlock Margin="16" Text="{DynamicResource ReloadOnWindowOpen}"/>
-                <ToggleButton Style="{StaticResource MaterialDesignSwitchToggleButton}" IsChecked="{Binding Reload, UpdateSourceTrigger=PropertyChanged}" ToolTip="{DynamicResource ReloadOnWindowOpen}" />
-            </StackPanel>
-
-            <Button Grid.Row="3" Margin="16" IsDefault="True" IsCancel="True" Style="{DynamicResource MaterialDesignFlatButton}" Command="{Binding CloseDialogCommand}" Content="{DynamicResource CloseSettings}"/>
-
-        </Grid>
-    </materialDesign:Card>
-</UserControl>
