--- TheGate/ViewModel/RelayCommand.cs
@@ -1,49 +1,0 @@
-﻿using System;
-using System.Diagnostics;
-using System.Windows.Input;
-
-namespace TheGate.ViewModel
-{
-    public class RelayCommand<T> : ICommand
-    {
-        readonly Action<T> _execute;
-        readonly Predicate<T> _canExecute;
-
-        public RelayCommand(Action<T> execute) : this(execute, null)
-        {
-        }
-
-        public RelayCommand(Action<T> execute, Predicate<T> canExecute)
-        {
-            if (execute == null)
-                throw new ArgumentNullException("execute");
-
-            _execute = execute;
-            _canExecute = canExecute;
-        }
-
-        [DebuggerStepThrough]
-        public bool CanExecute(object parameter)
-        {
-            return _canExecute == null ? true : _canExecute((T)parameter);
-        }
-
-        public event EventHandler CanExecuteChanged
-        {
-            add { CommandManager.RequerySuggested += value; }
-            remove { CommandManager.RequerySuggested -= value; }
-        }
-
-        public void Execute(object parameter)
-        {
-            _execute((T)parameter);
-        }
-    }
-
-    public class RelayCommand : RelayCommand<object>
-    {
-        public RelayCommand(Action execute) : this(execute, () => true) { }
-
-        public RelayCommand(Action execute, Func<bool> canExecute) : base(param => execute(), param => canExecute()) { }
-    }
-}
