--- TheGate/ViewModel/SettingBindingExtension.cs
@@ -1,23 +1,0 @@
-﻿using System.Windows.Data;
-
-namespace TheGate.ViewModel
-{
-    public class SettingBindingExtension : Binding
-    {
-        public SettingBindingExtension()
-        {
-            Initialize();
-        }
-
-        public SettingBindingExtension(string path) : base(path)
-        {
-            Initialize();
-        }
-
-        private void Initialize()
-        {
-            Source = Properties.Settings.Default;
-            Mode = BindingMode.TwoWay;
-        }
-    }
-}
